continue_rebase.go continue_rebase_if_needed.go
